refactor(workflow): decouple start node external dependencies
