+++ api/core/workflow/nodes/start/entities.py
@@ -2,8 +2,8 @@ from collections.abc import Sequence
 
 from pydantic import Field
 
-from core.app.app_config.entities import VariableEntity
 from core.workflow.nodes.base import BaseNodeData
+from core.workflow.variables.input_entities import VariableEntity
 
 
 class StartNodeData(BaseNodeData):
